+++ uxds_krb5.h
@@ -1,7 +1,7 @@
 /* krb.h */
 
 /* get short principal name from krb5 creds cache */
-char *get_krbname(authzdata auth, short parse);
+char *get_krbname(authzdata auth, int parse);
 /* lazy kinit */
 int get_tkts(char *user, char *service, authzdata auth);
 /* set password krb5Key */
